move other bit slicing gadget to bit slicing file
